Change library for Commandline parsing from NDesk to CommandLineParse… (#59)
Change from NDesk to CommandLineParser.
The implementation supports the original interface.
Except -? is now --help and the generated help text differs of course.

Because the CommandLineParser supports only a single minus followes
by a single character as shortCut, I was forced to add --ip and --IndexPath
als separate options. 

This is nasty but I wanted to support the old interface at the first approach.

I think after a cleaning up the interface of ICmdLineHandler, I would implement
a breaking cli as proposable.

Never the less now NDesk is not longer needed.